fixed bugs with nr_subpages_per_block; storage space becomes full with 128KB pages -- it might be not a problem with FTL implementation, so I expect that it would be solved with some optimizations

+++ ftl/algo/abm.h
@@ -67,7 +67,7 @@ typedef struct {
 	uint64_t chip_no;
 	uint64_t block_no;
 	uint32_t erase_count;
-	uint8_t nr_invalid_subpages;
+	uint32_t nr_invalid_subpages;
 	babm_abm_subpage_t* pst;	/* a page status table; used when the FTL requires */
 
 	struct list_head list;	/* for list */
